Dodavanje repository-ja kako bih mogla da dodam podatke preko Database Configuration jer sa data.sql nije nesto radilo

+++ backend/src/main/java/com/in51_2021/vebProjekat/entity/Recenzija.java
@@ -31,7 +31,7 @@ public class Recenzija {
     }
 
 
-    public Recenzija(Long ocena, String tekst,Date datumRecenzije) {
+    public Recenzija(int ocena, String tekst, Date datumRecenzije) {
         this.ocena = ocena;
         this.tekst = tekst;
         this.datumRecenzije = datumRecenzije;
